Ouch! Never upgrade in the past!
This solves the cause of many fatal problems experienced with upgrade process after an usage of the touch command on the repository.
Touch command is exeptionally used to force a vigiupdate.sh reboot or rebuild on all robots.
Because git pull refresh all files on the workdir even not modified one, touched file do a jump in the past!